refactor(gsd): extract auto-loop phase modules from auto/phases.ts
- Extract runPreDispatch to auto/pre-dispatch.ts
- Extract runDispatch to auto/dispatch.ts
- Extract runUnitPhase to auto/unit-phase.ts
- Extract runFinalize to auto/finalize.ts
- Extract closeout/merge helpers to auto/closeout.ts
- Extract shared phase helpers to auto/phase-helpers.ts
- Extract worktree-safety phase helpers to auto/worktree-safety-phase.ts
- Reduce auto/phases.ts to a compatibility shim re-exporting from new modules
- Update loop.ts, workflow-unit-dispatch.ts, and test imports
- Add new decision-path modules to parsers-legacy-importers banned list

Behavior-neutral refactor; all focused auto-loop tests pass.
Full suite: 10707 passed, 8 pre-existing/environmental failures.
