+++ src/resources/extensions/gsd/tests/parsers-legacy-importers.test.ts
@@ -28,6 +28,11 @@ const BANNED_DECISION_PATHS = new Set([
   "gsd/auto-post-unit.ts",
   "gsd/milestone-closeout.ts",
   "gsd/auto/phases.ts",
+  "gsd/auto/pre-dispatch.ts",
+  "gsd/auto/dispatch.ts",
+  "gsd/auto/unit-phase.ts",
+  "gsd/auto/finalize.ts",
+  "gsd/auto/closeout.ts",
   "gsd/auto/orchestrator.ts",
   "gsd/auto/loop.ts",
   "gsd/tools/complete-slice.ts",
