+++ src/resources/extensions/gsd/tests/phases-merge-error-stops-auto.test.ts
@@ -10,7 +10,7 @@
  */
 
 import { createTestContext } from "./test-helpers.ts";
-import { runPreDispatch } from "../auto/phases.ts";
+import { runPreDispatch } from "../auto/pre-dispatch.ts";
 
 const { assertTrue, report } = createTestContext();
