+++ src/resources/extensions/gsd/tests/plan-gate-failed-doctor-heal-hint.test.ts
@@ -18,12 +18,12 @@ function readSrc(file: string): string {
   return readFileSync(join(gsdDir, file), "utf-8");
 }
 
-test("#4620: auto/phases plan gate failed message includes doctor heal hint", () => {
-  const src = readSrc("auto/phases.ts");
+test("#4620: auto/pre-dispatch plan gate failed message includes doctor heal hint", () => {
+  const src = readSrc("auto/pre-dispatch.ts");
   assert.match(
     src,
     /Plan gate failed-closed:[\s\S]*\/gsd doctor heal/,
-    "auto/phases.ts should include /gsd doctor heal in plan gate failed notification",
+    "auto/pre-dispatch.ts should include /gsd doctor heal in plan gate failed notification",
   );
 });
